requirements.txt
python-telegram-bot==20.3
flask==2.3.2
openai
gunicorn
